abstract api_request() with custom rate-limit headers
also moves has_more_pages() into rest_api_client.py module
